requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-generator==1.10
async-timeout==4.0.3
attrs==21.4.0
azure-core==1.32.0
beautifulsoup4==4.12.3
blinker==1.8.2
botbuilder-core==4.16.2
botbuilder-integration-aiohttp==4.16.2
botbuilder-schema==4.16.2
botframework-connector==4.16.2
botframework-streaming==4.16.2
boto3==1.37.38
botocore==1.37.38
cachetools==5.5.0
certifi==2022.5.18.1
cffi==1.15.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
cryptography==37.0.2
dataclasses-json==0.6.7
deepgram-sdk==3.8.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
dotenv==0.9.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-speech==2.30.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.63.0rc1
h11==0.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.3
immutabledict==4.2.0
importlib_metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonpickle==1.4.2
MarkupSafe==2.1.5
marshmallow==3.22.0
meraki==1.53.0
mpmath==1.3.0
msal==1.31.1
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
openai==0.28.0
openpyxl==3.1.5
ortools==9.11.4210
outcome==1.1.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
pyhumps==3.8.0
PyJWT==2.9.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.11.5
safetensors==0.5.2
selenium==4.1.5
six==1.17.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.6
srvlookup==3.0.0
sympy==1.13.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
trio==0.20.0
trio-websocket==0.9.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.9
vboxapi==1.0
websockets==13.1
Werkzeug==3.0.6
wsproto==1.1.0
wxcadm==4.4.1
xlrd==2.0.1
xmltodict==0.14.2
yarl==1.15.2
zipp==3.20.2
